add template image during creation

+++ infra/kvm/bin/create_vm.sh
@@ -9,6 +9,12 @@ else
     exit 1
 fi
 
-virsh vol-delete $TARGETFOLDER/$NAME.img
-virsh vol-clone --pool $TARGETPOOL trusty-server-cloudimg-amd64-disk1.img $NAME.img
+if virsh vol-list --pool "$TARGETPOOL" | grep $TARGETFOLDER/$NAME.img > /dev/null; then
+    virsh vol-delete $TARGETFOLDER/$NAME.img
+    if virsh vol-list --pool "$TARGETPOOL" | grep $TARGETFOLDER/template.img > /dev/null; then
+        virsh vol-clone --pool $TARGETPOOL template.img $NAME.img
+    else
+        virsh vol-clone --pool $TARGETPOOL trusty-server-cloudimg-amd64-disk1.img $NAME.img
+    fi
+fi
 virsh create $NAME.xml --console
